requirements.txt
streamlit>=1.25.0
numpy>=1.25.0
opencv-python>=4.8.0
Pillow>=10.0.0
imageio>=2.31.0
moviepy>=1.0.0
